requirements.txt
python-dotenv==1.2.2
streamlit==1.55.0
nltk==3.9.4